feat(mixin, new decorator): mixin refactored to not be used as a function. Order key decorator
The ordered mixin was refactored to grab the tablename automatically, not needing it to be passed
into through a mixin function call. The order key was also removed in favor of a decorator. The
readme and tests also received significant updates.

BREAKING CHANGE: Ordered mixin contract was updated and the orderKey decorator now dictates tracking
seperate sequences on the same table.